Removed static from search method
Removed static from the search method in DroneFleet, added a drone
status contructor that takes in a drone fleet object as a parameter. The
original orientation of the search method in drone status meant that
divergent changes were made in the drone fleet class, this was a form of
change prevention.