New Method in the Drone Class to find availableDrone
Created new method in Drone class called availableDrone. The method takes two arguments, drones(List<Drone>), weight.
This  abides by the principles of oop, it makes the code more modular which allows other parts of the program to use it. It also
reduces the amount of clutter within the main. In theory the Drone class could be taken from this project and used in
another project. This abides by the DRY(Don't Repeat Yourself) principle of programming in object-oriented way.